Fix compilation warnings.
Conflicts:
	src/leaf-node-xyzaxis.cpp
	src/node.cpp

+++ src/node.cpp
@@ -27,7 +27,7 @@ namespace graphics {
       cylinder->addDrawable (new osg::ShapeDrawable (new osg::Cylinder (osg::Vec3(0,0,0), 1*factor, 4*factor)));
       group->addChild (cylinder);
       ::osg::GeodeRefPtr cone = new osg::Geode;
-      cone->addDrawable (new osg::ShapeDrawable (new osg::Cone (osg::Vec3(0.f,0.f,0.f), 1.8f*factor, 2*factor)));
+      cone->addDrawable (new osg::ShapeDrawable (new osg::Cone (osg::Vec3(0,0,0), 1.8f*factor, 2.0f*factor)));
       ::osg::PositionAttitudeTransformRefPtr cone_transform = new ::osg::PositionAttitudeTransform;
       cone_transform->addChild (cone);
       cone_transform->setPosition (osg::Vec3 (0,0,2*factor));
@@ -38,7 +38,7 @@ namespace graphics {
       const float radius = 1; const int POLYGON_SIZE = 256;
       float theta, px, py;
       for (int i = 1; i <= POLYGON_SIZE; ++i) {
-        theta = 2. * osg::PI  / POLYGON_SIZE * i;
+        theta = 2.f * (float)osg::PI / (float)(POLYGON_SIZE * i);
         px = radius * cosf (theta);
         py = radius * sinf (theta);
         v->push_back (osg::Vec3(px,py,0));
